move file location. hang intentionally when allocation failed
